requirements.txt
streamlit==1.32.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.2.0
PyPDF2==3.0.1
python-docx==1.1.0
